requirements.txt
nltk==3.8.1
transformers==4.36.2
torch>=2.0.0
pyyaml==6.0.1
tqdm
